pyproject.toml
[project]
name = "omc-pt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "pydantic[email]>=2.13.3",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "psycopg[binary]>=3.2.12",
    "uvicorn>=0.46.0",
    "pydantic-settings>=2.14.0",
    "google-genai>=1.73.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
]
